refactor(runner): split scheduler dependency index roles
